Refactor package structure for better readability
- Moved all "reader" types from `util` to `reader` pkg
- Moved all "model" types from `util` and `formats` to `model` pkg
- Moved all "iter" types from `util` to `iter` pkg
