+++ src/main/java/de/digitalcollections/solrocr/formats/alto/AltoPassageFormatter.java
@@ -1,10 +1,10 @@
 package de.digitalcollections.solrocr.formats.alto;
 
 import de.digitalcollections.solrocr.formats.OcrPassageFormatter;
-import de.digitalcollections.solrocr.util.IterableCharSequence;
-import de.digitalcollections.solrocr.util.OcrBox;
-import de.digitalcollections.solrocr.util.OcrPage;
-import de.digitalcollections.solrocr.util.TagBreakIterator;
+import de.digitalcollections.solrocr.iter.IterableCharSequence;
+import de.digitalcollections.solrocr.model.OcrBox;
+import de.digitalcollections.solrocr.model.OcrPage;
+import de.digitalcollections.solrocr.iter.TagBreakIterator;
 import java.awt.Dimension;
 import java.util.ArrayList;
 import java.util.HashMap;
