+++ src/main/java/de/digitalcollections/solrocr/iter/ExitingIterCharSeq.java
@@ -1,5 +1,6 @@
-package de.digitalcollections.solrocr.util;
+package de.digitalcollections.solrocr.iter;
 
+import de.digitalcollections.solrocr.model.SourcePointer;
 import java.nio.charset.Charset;
 import java.util.stream.IntStream;
 import org.apache.lucene.index.QueryTimeout;
