+++ src/main/java/de/digitalcollections/solrocr/lucene/OcrHighlighter.java
@@ -1,22 +1,22 @@
 package de.digitalcollections.solrocr.lucene;
 
 import com.google.common.collect.ImmutableSet;
-import de.digitalcollections.solrocr.formats.OcrBlock;
-import de.digitalcollections.solrocr.formats.OcrFormat;
+import de.digitalcollections.solrocr.model.OcrBlock;
+import de.digitalcollections.solrocr.model.OcrFormat;
 import de.digitalcollections.solrocr.formats.OcrPassageFormatter;
-import de.digitalcollections.solrocr.formats.OcrSnippet;
+import de.digitalcollections.solrocr.model.OcrSnippet;
 import de.digitalcollections.solrocr.formats.alto.AltoFormat;
 import de.digitalcollections.solrocr.formats.hocr.HocrFormat;
 import de.digitalcollections.solrocr.formats.mini.MiniOcrFormat;
 import de.digitalcollections.solrocr.solr.OcrHighlightParams;
-import de.digitalcollections.solrocr.util.ExitingIterCharSeq;
-import de.digitalcollections.solrocr.util.FileBytesCharIterator;
+import de.digitalcollections.solrocr.iter.ExitingIterCharSeq;
+import de.digitalcollections.solrocr.iter.FileBytesCharIterator;
 import de.digitalcollections.solrocr.util.HighlightTimeout;
-import de.digitalcollections.solrocr.util.IterableCharSequence;
-import de.digitalcollections.solrocr.util.MultiFileBytesCharIterator;
-import de.digitalcollections.solrocr.util.OcrHighlightResult;
+import de.digitalcollections.solrocr.iter.IterableCharSequence;
+import de.digitalcollections.solrocr.iter.MultiFileBytesCharIterator;
+import de.digitalcollections.solrocr.model.OcrHighlightResult;
 import de.digitalcollections.solrocr.util.PageCacheWarmer;
-import de.digitalcollections.solrocr.util.SourcePointer;
+import de.digitalcollections.solrocr.model.SourcePointer;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
