+++ src/main/java/de/digitalcollections/solrocr/model/OcrBlock.java
@@ -1,4 +1,4 @@
-package de.digitalcollections.solrocr.formats;
+package de.digitalcollections.solrocr.model;
 
 public enum OcrBlock {
   /* Order matters: From top of the page layout hierarchy to bottom */
