+++ src/test/java/de/digitalcollections/solrocr/lucene/ExternalUtf8ContentFilterTest.java
@@ -4,8 +4,8 @@ package de.digitalcollections.solrocr.lucene;
 import com.google.common.collect.ImmutableList;
 import de.digitalcollections.solrocr.lucene.filters.ExternalUtf8ContentFilter;
 import de.digitalcollections.solrocr.lucene.filters.ExternalUtf8ContentFilterFactory;
-import de.digitalcollections.solrocr.util.MultiFileReader;
-import de.digitalcollections.solrocr.util.SourcePointer;
+import de.digitalcollections.solrocr.reader.MultiFileReader;
+import de.digitalcollections.solrocr.model.SourcePointer;
 import de.digitalcollections.solrocr.util.Utf8;
 import java.io.BufferedReader;
 import java.io.FileReader;
