+++ src/test/java/de/digitalcollections/solrocr/util/MultiFileBytesCharIteratorTest.java
@@ -3,6 +3,7 @@ package de.digitalcollections.solrocr.util;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.google.common.collect.ImmutableList;
+import de.digitalcollections.solrocr.iter.MultiFileBytesCharIterator;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
