+++ src/test/java/de/digitalcollections/solrocr/util/PeekingReaderTest.java
@@ -1,5 +1,6 @@
 package de.digitalcollections.solrocr.util;
 
+import de.digitalcollections.solrocr.reader.PeekingReader;
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.Path;
