+++ src/test/java/de/digitalcollections/solrocr/util/TagBreakIteratorTest.java
@@ -3,6 +3,9 @@ package de.digitalcollections.solrocr.util;
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
 import com.google.common.collect.ImmutableSet;
+import de.digitalcollections.solrocr.iter.FileBytesCharIterator;
+import de.digitalcollections.solrocr.iter.IterableCharSequence;
+import de.digitalcollections.solrocr.iter.TagBreakIterator;
 import java.io.IOException;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
